app/requirements.txt
fastapi=0.88.0
numpy=1.23.4
pandas=1.5.2
pymongo=4.3.3
python-dotenv=0.21.0
spacy=3.4.4
torch=1.13.1
uvicorn=0.20.0
bcrypt=4.0.1
httpx=0.23.3
PyJWT=2.6.0
twilio-7.16.1